refactor(deferred): move reference precompiles to integration tests; add in-lib engine fixture
Step 1 of the precompile-framework redesign. NOT a pure file move — a
test-architecture refactor. The original plan's "pure move" premise was wrong:
Uint256 was also the schema vehicle for miden-core's own state.rs unit tests
and for processor/tests/deferred_e2e.rs, neither of which the plan addressed.

- Reference apps Uint256/MockGroup/MockHash/MockSig move out of miden-core's
  lib (core/src/deferred/app/) into integration-test-private
  core/tests/common/precompile/ as Uint/Group/Hash/Sig (impl only). FieldOps
  travels with Uint. The testing-gated re-exports are removed from
  core/src/deferred/{mod,app/mod}.rs: no reference impls on miden-core's public
  or `testing` surface in any configuration.
- core/tests/deferred_{uint256,mock_*}.rs renamed to
  core/tests/precompile_{uint,group,hash,sig}.rs, rewired via `mod common;`,
  with the reference impls' former in-lib unit tests refolded here against the
  public API (private-structure-only cases dropped, covered via public decode).
- New core/src/deferred/test_precompile.rs: a #[cfg(test)] pub(crate) minimal
  schema fixture (value/add/mul/eq). state.rs's engine unit tests use it
  instead of borrowing Uint256 — engine tests should depend on the simplest
  schema, not a reference app.
- processor/tests/deferred_e2e.rs: replaces the Uint256 vehicle with an in-file
  ArithTestSchema (mirrors the existing local ChunkTestSchema idiom); no
  cross-crate reference-impl coupling. Uint::init now registers its constants
  via the public Schema API (DeferredState::intern is crate-private).

Production semantics unchanged; only test scaffolding moved. Verified:
cargo check --workspace --all-targets clean (no warnings); miden-core lib (366)
+ precompile_{uint,group,hash,sig} + deferred_chunk + processor deferred_e2e
all green.
